+++ packages/node/CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
+### Added
+- `store-cache-upper-limit` flag to limit the max size of the store cache. (#124)
 
 ## [2.8.0] - 2023-06-27
 ### Changed